+++ cli/device.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/mobile-next/mobilecli/commands"
-	"github.com/mobile-next/mobilecli/devices"
 	"github.com/spf13/cobra"
 )
 
@@ -38,23 +37,6 @@ var deviceInfoCmd = &cobra.Command{
 	Short: "Get device info",
 	Long:  `Get detailed information about a connected device, such as OS, version, and screen size.`,
 	RunE: func(cmd *cobra.Command, args []string) error {
-
-		// Find the target device
-		targetDevice, err := commands.FindDeviceOrAutoSelect(deviceId)
-		if err != nil {
-			response := commands.NewErrorResponse(fmt.Errorf("error finding device: %v", err))
-			printJson(response)
-			return fmt.Errorf("%s", response.Error)
-		}
-
-		// Start agent
-		err = targetDevice.StartAgent(devices.StartAgentConfig{})
-		if err != nil {
-			response := commands.NewErrorResponse(fmt.Errorf("error starting agent: %v", err))
-			printJson(response)
-			return fmt.Errorf("%s", response.Error)
-		}
-
 		response := commands.InfoCommand(deviceId)
 		printJson(response)
 		if response.Status == "error" {