+++ commands/orientation.go
@@ -30,7 +30,9 @@ func OrientationGetCommand(req OrientationGetRequest) *CommandResponse {
 	}
 
 	// start agent if needed
-	err = device.StartAgent(devices.StartAgentConfig{})
+	err = device.StartAgent(devices.StartAgentConfig{
+		Hook: GetShutdownHook(),
+	})
 	if err != nil {
 		return NewErrorResponse(fmt.Errorf("failed to start agent on device %s: %v", device.ID(), err))
 	}
@@ -60,7 +62,9 @@ func OrientationSetCommand(req OrientationSetRequest) *CommandResponse {
 	}
 
 	// start agent if needed
-	err = device.StartAgent(devices.StartAgentConfig{})
+	err = device.StartAgent(devices.StartAgentConfig{
+		Hook: GetShutdownHook(),
+	})
 	if err != nil {
 		return NewErrorResponse(fmt.Errorf("failed to start agent on device %s: %v", device.ID(), err))
 	}